Adding Animation for Crouching
There's idle crouching, which I think a single frame is enough, no need to make the character hump the ground. But you can add as many as you think are good.

There's also crouch walking, we can do more or less than 3.

I'm not sure if it's worth to go through with this feature, as it add little to the game, but may add a lot to the art part.

If we have time, it may be nice, but it's not a main matter, this can definitely be reverted.