Adding crouching animation functionality
It's bounded to the left shift key as something you have to hold. I may add a toggle option later.

Regarding the animation, it player "Walk (Crouched)" while walking on the floor and "Idle (Crouched)" for all other cases, even while on air. Mario does this and it should save the artists a good bit of time.